Переносит всю логику по работе с редактором из RoutePresenter в EditorPresenter
