--- src/model/route-model-types.js
@@ -1,21 +1,0 @@
-/**
- * @typedef AggregatedPoint
- * @prop {number} basePrice
- * @prop {string} dateFrom
- * @prop {string} dateTo
- * @prop {Destination} destination
- * @prop {string} id
- * @prop {Offer[]} offers
- * @prop {PointType} type
- */
-
-/**
- * @typedef PointAdapter
- * @prop {number} basePrice
- * @prop {string} startDate
- * @prop {string} endDate
- * @prop {number} destinationId
- * @prop {string} id
- * @prop {number[]} offerIds
- * @prop {PointType} type
- */
