+++ src/model/route-model.js
@@ -11,19 +11,13 @@ import PointAdapter from '../adapter/point-adapter.js';
 const clone = (target) => JSON.parse(JSON.stringify(target));
 
 export default class RouteModel extends EventTarget {
-  /**
-   * @type {Point[]}
-   */
+  /** @type {Point[]} */
   #points = null;
 
-  /**
-   * @type {Destination[]}
-   */
+  /** @type {Destination[]} */
   #destinations = null;
 
-  /**
-   * @type {OfferGroup[]}
-   */
+  /** @type {OfferGroup[]} */
   #offerGroups = null;
 
   async ready() {
@@ -39,16 +33,14 @@ export default class RouteModel extends EventTarget {
   }
 
   getPoints() {
-    return this.#points.map((point) => new PointAdapter(point));
+    return this.#points.map((point) => new PointAdapter(clone(point)));
   }
 
   /**
-   * @param {string} id
+   * @param {number} id
    */
   getPointById(id) {
-    const point = this.#points.find((item) => (item.id === id));
-
-    return clone(new PointAdapter(point));
+    return this.getPoints().find((item) => (item.id === id));
   }
 
   getOfferGroups() {
