+++ src/presenter/route-presenter.js
@@ -1,41 +1,37 @@
 /** @typedef {import('../model/route-model').default} RouteModel */
 /** @typedef {import('../view/editor-view').default} EditorView */
+/** @typedef {import('../adapter/point-adapter').default} PointAdapter */
 
 import RouteView from '../view/route-view.js';
 import PointView from '../view/point-view.js';
-import Type from '../enum/type.js';
-import TypeLabel from '../enum/type-label.js';
-import { formatDate, getOfferSelectOptions } from '../utils.js';
+import { formatDate } from '../utils.js';
 
 export default class RoutePresenter {
   /**
    * @param {RouteModel} model
    * @param {EditorView} model
    */
-  constructor(model, editorView) {
+  constructor(model) {
     this.model = model;
 
     /**
      * @type {RouteView}
      */
     this.view = document.querySelector(String(RouteView));
-    this.editorView = editorView;
 
-    this.model.ready().then(() => {
-      const points = this.model.getPoints();
-      const pointViews = points.map((point) => this.createPointView(point));
-      const isRouteEmpty = (points.length === 0);
+    const points = this.model.getPoints();
+    const pointViews = points.map((point) => this.createPointView(point));
+    const isRouteEmpty = (points.length === 0);
 
-      if (isRouteEmpty) {
-        this.view.showPlaceholder('Click New Event to create your first point');
+    if (isRouteEmpty) {
+      this.view.showPlaceholder('Click New Event to create your first point');
 
-        return;
-      }
+      return;
+    }
 
-      this.view
-        .hidePlaceholder()
-        .setPoints(...pointViews);
-    });
+    this.view
+      .hidePlaceholder()
+      .setPoints(...pointViews);
   }
 
   get dateFormat() {
@@ -54,7 +50,7 @@ export default class RoutePresenter {
    * @param {PointAdapter} point
    */
   createPointView(point) {
-    const view = new PointView();
+    const view = new PointView(point.id);
     const destination = this.model.getDestinationById(point.destinationId);
     const title = `${point.type} ${destination.name}`;
     const price = String(point.basePrice);
@@ -78,94 +74,6 @@ export default class RoutePresenter {
 
     view.pointOffersView.setOptions(offersOptions);
 
-    view.addEventListener('expand', () => {
-      this.editorView.close();
-      this.updatePointEditorView(point);
-      this.editorView
-        .link(view)
-        .open();
-    });
-
     return view;
-  }
-
-  /**
-   * @param {PointAdapter} point
-   */
-  updatePointEditorView(point) {
-    const typeTitle = TypeLabel[Type.findKey(point.type)];
-    const destination = this.model.getDestinationById(point.destinationId);
-
-    const startDate = formatDate(point.startDate, this.dateFormat);
-    const startTime = formatDate(point.startDate, this.timeFormat);
-    const startDateTime = `${startDate} ${startTime}`;
-
-    const endDate = formatDate(point.endDate, this.dateFormat);
-    const endTime = formatDate(point.endDate, this.timeFormat);
-    const endDateTime = `${endDate} ${endTime}`;
-
-    /**
-     * @type {[string, PointType, boolean][]}
-     */
-    const typeSelectOptions = Object.values(Type).map((type) => {
-      const key = Type.findKey(type);
-      const label = TypeLabel[key];
-      const isChecked = (type === point.type);
-
-      return [label, type, isChecked];
-    });
-
-    /**
-     * @type {[string, string][]}
-     */
-    const destinationInputOptions = this.model.getDestinations().map(
-      (item) => ['', item.name]
-    );
-
-    /**
-     * @type {[number, string, number, boolean][]}
-     */
-    const offerSelectOptions = getOfferSelectOptions(
-      this.model.getAvailableOffers(point.type),
-      point.offerIds
-    );
-
-    /**
-     * @type {[string, string][]}
-     */
-    const pictureOptions = destination.pictures.map(
-      ({ src, description }) => [ src, description ]
-    );
-
-    const {
-      typeSelectView,
-      destinationInputView,
-      datePickerView,
-      priceInputView,
-      offerSelectView,
-      destinationDetailsView
-    } = this.editorView;
-
-    typeSelectView
-      .setIcon(point.type)
-      .setOptions(typeSelectOptions);
-
-    destinationInputView
-      .setLabel(typeTitle)
-      .setValue(destination.name)
-      .setOptions(destinationInputOptions);
-
-    datePickerView
-      .setStartDate(startDateTime)
-      .setEndDate(endDateTime);
-
-    priceInputView.setValue(point.basePrice);
-    offerSelectView.setOptions(offerSelectOptions);
-
-    destinationDetailsView
-      .setDescription(destination.description)
-      .setPictures(pictureOptions);
-
-    return this.editorView;
   }
 }
